requirements.txt
agent-framework-core>=1.0.0rc3
agent-framework-ag-ui>=1.0.0b260304
agent-framework-devui>=1.0.0b260304
aiohttp>=3.10.0
azure-identity>=1.17.1
fastapi>=0.111.0
httpx>=0.27.0
jinja2>=3.1.4
numpy>=1.26.4
openai>=1.40.0
pandas>=2.2.2
plotly>=5.22.0
python-dotenv>=1.0.1
requests>=2.32.3
rich>=13.7.1
scipy>=1.13.1
streamlit>=1.36.0
tavily-python>=0.5.0
uvicorn[standard]>=0.30.1
websockets>=12.0